Backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
brotlicffi==1.1.0.0
celery==5.5.3
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
Django==4.2.16
django-axes==6.1.0
django-cors-headers==4.3.1
django-cryptography
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
kombu==5.5.4
msgpack==1.1.0
olefile==0.46
packaging==25.0
pandas==2.2.2
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
redis==5.0.8
service-identity==24.2.0
setuptools<81
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.0.0
zope.interface==7.2
